requirements.dev.txt
flake8>=6.1.0,<6.2
pytest_django>=4.7.0,<4.8
factory_boy>=3.3.0,<3.4
debugpy>=1.8.0,<1.9